+++ src/index.ts
@@ -1,27 +1,36 @@
 import { isRef } from "./ref"
-import { isString, isArrayLike, isBoolean, isObject, isFunction, isElement, isNumber, keys } from "./util"
+import {
+  isString,
+  isArrayLike,
+  isBoolean,
+  isObject,
+  isFunction,
+  isElement,
+  isNumber,
+  keys,
+} from "./util"
 
 export { createRef } from "./ref"
 
-declare const __SVG__: boolean;
+declare const __SVG__: boolean
 
 export const SVGNamespace = "http://www.w3.org/2000/svg"
 const XLinkNamespace = "http://www.w3.org/1999/xlink"
 const XMLNamespace = "http://www.w3.org/XML/1998/namespace"
 
 export function preventDefault(event: Event) {
-  event.preventDefault();
-  return event;
+  event.preventDefault()
+  return event
 }
 export function stopPropagation(event: Event) {
-  event.stopPropagation();
-  return event;
+  event.stopPropagation()
+  return event
 }
 
 // https://facebook.github.io/react/docs/jsx-in-depth.html#booleans-null-and-undefined-are-ignored
 // Emulate JSX Expression logic to ignore certain type of children or className.
 function isVisibleChild(value: any): boolean {
-  return !isBoolean(value) && value != null;
+  return !isBoolean(value) && value != null
 }
 
 /**
@@ -30,13 +39,18 @@ function isVisibleChild(value: any): boolean {
  */
 function className(value: any): string {
   if (Array.isArray(value)) {
-    return value.map(className).filter(Boolean).join(" ");
+    return value
+      .map(className)
+      .filter(Boolean)
+      .join(" ")
   } else if (isObject(value)) {
-    return keys(value).filter(k => value[k]).join(" ");
+    return keys(value)
+      .filter(k => value[k])
+      .join(" ")
   } else if (isVisibleChild(value)) {
-    return "" + value;
+    return "" + value
   } else {
-    return "";
+    return ""
   }
 }
 
@@ -95,66 +109,70 @@ const svg = {
   tspan: 0,
   use: 0,
   view: 0,
-};
-
-export function Fragment(attr: { children: JSX.Child[] }) {
-  const fragment = document.createDocumentFragment();
-  appendChildren(attr.children, fragment);
-  return fragment;
 }
 
-export { createElement as h };
+export function createFactory(tag: string) {
+  return createElement.bind(null, tag)
+}
+
+export function Fragment(attr: { children: JSX.Child[] }) {
+  const fragment = document.createDocumentFragment()
+  appendChildren(attr.children, fragment)
+  return fragment
+}
+
+export { createElement as h }
 export function createElement(tag, attr, ...children) {
-  attr = attr || {};
+  attr = attr || {}
   if (__SVG__ && !attr.namespaceURI && svg[tag] === 0) {
-    attr = { ...attr, namespaceURI: SVGNamespace };
+    attr = { ...attr, namespaceURI: SVGNamespace }
   }
 
-  let node: HTMLElement | SVGElement;
+  let node: HTMLElement | SVGElement
   if (isString(tag)) {
     node = attr.namespaceURI
       ? document.createElementNS(attr.namespaceURI, tag)
-      : document.createElement(tag);
-    attributes(attr, node);
-    appendChildren(children, node);
+      : document.createElement(tag)
+    attributes(attr, node)
+    appendChildren(children, node)
   } else if (isFunction(tag)) {
     // Custom elements.
     if (isObject(tag.defaultProps)) {
-      attr = { ...tag.defaultProps, ...attr };
+      attr = { ...tag.defaultProps, ...attr }
     }
 
-    node = tag({ ...attr, children });
+    node = tag({ ...attr, children })
   }
 
   if (isRef(attr.ref)) {
-    attr.ref.current = node;
+    attr.ref.current = node
   } else if (isFunction(attr.ref)) {
-    attr.ref(node);
+    attr.ref(node)
   }
-  return node;
+  return node
 }
 
 function appendChild(child, node: Node) {
   if (isArrayLike(child)) {
-    appendChildren(child, node);
+    appendChildren(child, node)
   } else if (isString(child) || isNumber(child)) {
-    node.appendChild(document.createTextNode(child as any));
+    node.appendChild(document.createTextNode(child as any))
   } else if (child === null) {
-    node.appendChild(document.createComment(""));
+    node.appendChild(document.createComment(""))
   } else if (isElement(child)) {
-    node.appendChild(child);
+    node.appendChild(child)
   }
 }
 
 function appendChildren(children, node: Node) {
   for (const child of children) {
-    appendChild(child, node);
+    appendChild(child, node)
   }
-  return node;
+  return node
 }
 
 function normalizeAttribute(s: string) {
-  return s.replace(/[A-Z\d]/g, match => ":" + match.toLowerCase());
+  return s.replace(/[A-Z\d]/g, match => ":" + match.toLowerCase())
 }
 
 function attribute(key: string, value: any, node: HTMLElement | SVGElement) {
@@ -167,70 +185,69 @@ function attribute(key: string, value: any, node: HTMLElement | SVGElement) {
       case "xlinkShow":
       case "xlinkTitle":
       case "xlinkType":
-        node.setAttributeNS(XLinkNamespace, normalizeAttribute(key), value);
-        return;
+        node.setAttributeNS(XLinkNamespace, normalizeAttribute(key), value)
+        return
       case "xmlnsXlink":
-        node.setAttribute(normalizeAttribute(key), value);
-        return;
+        node.setAttribute(normalizeAttribute(key), value)
+        return
       case "xmlBase":
       case "xmlLang":
       case "xmlSpace":
-        node.setAttributeNS(XMLNamespace, normalizeAttribute(key), value);
-        return;
+        node.setAttributeNS(XMLNamespace, normalizeAttribute(key), value)
+        return
     }
   }
 
   switch (key) {
     case "htmlFor":
-      node.setAttribute("for", value);
-      return;
+      node.setAttribute("for", value)
+      return
     case "dataset":
       for (const dataKey of keys<object>(value || {})) {
-        const dataValue = value[dataKey];
+        const dataValue = value[dataKey]
         if (dataValue != null) {
-          (node as HTMLElement).dataset[dataKey] = dataValue;
+          ;(node as HTMLElement).dataset[dataKey] = dataValue
         }
       }
-      return;
+      return
     case "innerHTML":
     case "innerText":
     case "textContent":
-      node[key] = value;
-      return;
+      node[key] = value
+      return
     case "spellCheck":
-      (node as HTMLInputElement).spellcheck = value;
-      return;
+      ;(node as HTMLInputElement).spellcheck = value
+      return
     case "class":
     case "className":
-      node.setAttribute("class", className(value));
-      return;
+      node.setAttribute("class", className(value))
+      return
     case "ref":
     case "namespaceURI":
-      return;
+      return
     case "style":
       if (isObject(value)) {
-        Object.assign(node.style, value);
-        return;
+        Object.assign(node.style, value)
+        return
       }
-      // fallthrough
+    // fallthrough
   }
 
   if (isFunction(value)) {
     if (key[0] === "o" && key[1] === "n") {
-      const name = key.slice(2).toLowerCase();
-      node.addEventListener(name, value);
+      const name = key.slice(2).toLowerCase()
+      node.addEventListener(name, value)
     }
   } else if (value === true) {
-    node.setAttribute(key, "");
+    node.setAttribute(key, "")
   } else if (value !== false && value != null) {
-    node.setAttribute(key, value);
+    node.setAttribute(key, value)
   }
 }
 
 function attributes(attr: object, node: HTMLElement | SVGElement) {
   for (const key of keys(attr)) {
-    attribute(key, attr[key], node);
+    attribute(key, attr[key], node)
   }
-  return node;
+  return node
 }
-